+++ examples/next-ai-rsc/app/action.tsx
@@ -126,7 +126,11 @@ Messages inside [] means that it's a UI element or a user event. For example:
 - "[Price of AAPL = 100]" means that an interface of the stock price of AAPL is shown to the user.
 - "[User has changed the amount of AAPL to 10]" means that the user has changed the amount of AAPL to 10 in the UI.
 
-Always double confirm with the user before making an actual purchase.
+If the user requests purchasing a stock, call \`show_stock_purchase_ui\` to show the purchase UI.
+If the user just wants the price, call \`show_stock_price\` to show the price.
+If you want to show trending stocks, call \`list_stocks\`.
+If you want to show events, call \`get_events\`.
+If the user wants to sell stock, or complete another impossible task, respond that you are a demo and cannot do that.
 
 Besides that, you can also chat with users and do some calculations if needed.`,
       },
@@ -262,15 +266,7 @@ Besides that, you can also chat with users and do some calculations if needed.`,
 
   completion.onFunctionCall(
     'show_stock_price',
-    async ({
-      symbol,
-      price,
-      delta,
-    }: {
-      symbol: string;
-      price: number;
-      delta: number;
-    }) => {
+    async ({ symbol, price, delta }) => {
       reply.update(
         <BotCard>
           <StockSkeleton />
@@ -298,15 +294,7 @@ Besides that, you can also chat with users and do some calculations if needed.`,
 
   completion.onFunctionCall(
     'show_stock_purchase_ui',
-    ({
-      symbol,
-      price,
-      numberOfShares = 100,
-    }: {
-      symbol: string;
-      price: number;
-      numberOfShares?: number;
-    }) => {
+    ({ symbol, price, numberOfShares = 100 }) => {
       if (numberOfShares <= 0 || numberOfShares > 1000) {
         reply.done(<BotMessage>Invalid amount</BotMessage>);
         aiState.done([